New canonical Iterator
